requirements.txt
-r requirements-sys.txt
aenum; python_version < "3.6"
alembic>=1.3.0,<1.5; python_version <= "3.6"
alembic>=1.10; python_version >= "3.7"
# TODO: remove when merged
#   until fix merged and deployed (https://github.com/authomatic/authomatic/pull/195)
#   old variants:
#       -e git+https://github.com/fmigneault/authomatic.git@httplib-port#egg=Authomatic
#       https://github.com/fmigneault/authomatic/archive/httplib-port.zip#egg=Authomatic-1.0.1
#   new authomatic handles openid install correctly
#   leave http port until merged:
authomatic[OpenID] @ https://github.com/fmigneault/authomatic/archive/httplib-port.zip
bcrypt>=3.1.6
# FIXME: integrate when implemnted by official package (see https://github.com/bbangert/beaker/issues/201)
beaker @ https://github.com/crim-ca/beaker/archive/0ac88bcd8cca063a571fc385ffbe9bcc8acaa690.zip
colander
cornice<5; python_version < "3"
cornice; python_version >= "3"
cornice_swagger>=0.7.0
dicttoxml
# futures is required for gunicorn threads
futures; python_version < "3"
# flag --paste breaks for >20, must use pserve instead
gunicorn<20; python_version < "3.5"  # pyup: ignore
gunicorn>=20; python_version >= "3"
humanize
jsonschema<4; python_version < "3.6"
jsonschema>=4; python_version >= "3.6"
jwcrypto==1.5.0
lxml>=3.7
mako  # controlled by pyramid_mako
paste
pastedeploy
pluggy
psycopg2-binary>=2.7.1
PyJWT[crypto]==2.8.0
pyramid>=1.10.2,<2
pyramid_beaker==0.8
pyramid_chameleon>=0.3
pyramid_mako>=1.0.2
pyramid_retry==2.1.1
pyramid_tm>=2.2.1
python-dotenv
python2-secrets; python_version <= "3.5"
pyyaml>=5.1
requests
requests_file
simplejson
six>=1.12.0
# see also: safety 51668, https://github.com/sqlalchemy/sqlalchemy/pull/8563, still beta + major version change 2.0.0b1
sqlalchemy==1.3.*; python_version <= "3.6"  # pyup: ignore
sqlalchemy>=1.4.44,<2; python_version >= "3.7"  # pyup: ignore
# see also: safety 42194, https://github.com/kvesteri/sqlalchemy-utils/issues/166, not fixed since 2015
sqlalchemy-utils<0.36.4; python_version < "3"  # pyup: ignore
sqlalchemy-utils==0.37.9; python_version >= "3" and python_version <= "3.5"  # pyup: ignore
sqlalchemy-utils==0.39.0; python_version >= "3.6"  # pyup: ignore
threddsclient==0.4.2; python_version < "3"  # pyup: ignore
threddsclient>=0.4.2; python_version >= "3"
transaction
typing; python_version < "3"
# typing extension required for TypedDict
typing_extensions; python_version < "3.8"
wheel; python_version <= "3.6"
wheel>=0.38; python_version >= "3.7"
webob
ziggurat_foundations==0.8.4; python_version <= "3.6"  # pyup: ignore
ziggurat_foundations==0.9.1; python_version >= "3.7"
zope.interface>=4.7.2,<5
zope.sqlalchemy==1.6
